Fix max aliased point size reported by the driver
The graphics driver allows to render larger points than it reports to
support. This causes conformance tests to fail, because it is supposed
to clamp requested point sizes to what is reported to be supported. To
fix this, report a point size range that matches the actual behavior of
the driver.

Issue: FP2P-350
Test: cts -m CtsDeqpTestCases
  -t dEQP-GLES2.functional.rasterization.limits#points
Change-Id: I5721c37cf627e0c473b7070a0449a22ac759579e
(cherry picked from commit 13ae3089f9127f91a04296d9a59ac5258215e261)