libbinder: also avaid sWarningCallback lock
Since binders are made in FastNative or CriticalNative
code, they may be called while holding the mutator lock,
which another thread may try to lock, which already holds
the native lock held here.

This was previously fixed for calls to sLimitCallback, but
this fixes things for sWarningCallback as well, which was
added later.

Bug: 372506687
Bug: 354286280
Bug: 199683153
Bug: 352692435
Test: boot
Change-Id: Iafe1208b331df74ad2b2117ae0ceab4663f53c41